Found and fixed bug in MaterialLinearElastic4
The material computes the first Lamé constant λ and shear modulus µ in
`add_pixel`. Then in `compute_stress(_tangent)`, it is called with λ,
µ (as can be checked by looking into the order in which the fields are
arranged in the constructor), but the methods' declarations were
written expecting Young's E and Poisson's ratio ν, so they wrongly
recomputed λ(E,ν) and µ(E,ν) which then were totally off.