header tests now standalone compilable
tests that depend only on µSpectre headers and not on the library are
separated out of the main_test_suite. As a result, when developping
header-only features (such an all central data structures), one can
compile and run the corresponding tests (via make <test_name>) without
recompiling the entire library, thus significantly accelerating the
code-compile-test cycle for the central data structures and headers.