Globalisation of material-internal (local) fields now possible
Materials hold their held internal variables in local field
collections that they control. It is, however, very useful for
debugging and i/o to be able to access them like global fields.

Cells now have the ability to do exactly that: they create a global
field, and fill it with the values of the local fields, where they are available.